Update the use cases in the explainer's motivation section
This change updates the explainer to not list Prioritized Task
Scheduling as a browser-internal use case, since it can be used
together with `AsyncContext` as the former API is currently defined.
The browser-internal use case (transitive task prioritization) is also
listed, together with it.

This change also removes the browser-internal use case of limiting
access to certain APIs, since it was felt to be somewhat
controversial.