Add `CreateAsyncContextSnapshot` AO
In the web integration of AsyncContext, web specs will need to create
`AsyncContext.Snapshot` objects from an AsyncContext mapping; and we
expect this to be necessary for other kinds of hosts as well, such as
Node.js.

Currently the spec text does not provide any way to do that other than
manually creating the object and populating its
`[[AsyncContextMapping]]` slot, since the `AsyncContext.Snapshot`
constructor always uses the current mapping. So this patch adds an
abstract operation called `CreateAsyncContextSnapshot` that provides
this capability.